Separate brackets from listings
